# api/openapi.yaml
openapi: 3.0.3
info:
  title: Snake Game API
  description: |
    RESTful API for the Snake Game multiplayer application.
    Provides authentication, leaderboard, and game score management.
  version: 1.0.0
  contact:
    name: API Support
    email: support@snakegame.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.snakegame.com/api/v1
    description: Production server
...
paths:
...
  /leaderboard:
    get:
      tags:
        - Leaderboard
      summary: Get leaderboard
      description: Retrieve the top scores leaderboard
      operationId: getLeaderboard
      parameters:
        - name: limit
          in: query
          description: Maximum number of entries to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          example: 10
        - name: mode
          in: query
          description: Filter by game mode
          required: false
          schema:
            $ref: '#/components/schemas/GameMode'
      responses:
        '200':
          description: Leaderboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardResponse'
              example:
                success: true
                leaderboard:
                  - id: 3
                    userId: 1
                    username: player1
                    score: 200
                    mode: wall
                    timestamp: 1704067200000
                    rank: 1
                  - id: 1
                    userId: 1
                    username: player1
                    score: 150
                    mode: wall
                    timestamp: 1704063600000
                    rank: 2
                  - id: 2
                    userId: 2
                    username: player2
                    score: 120
                    mode: pass
                    timestamp: 1704056400000
                    rank: 3
  /scores:
    post:
      tags:
        - Leaderboard
      summary: Submit a game score
      description: Submit a new game score for the authenticated user
      operationId: submitScore
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreSubmission'
            example:
              score: 150
              mode: wall
      responses:
        '201':
          description: Score submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResponse'
              example:
                success: true
                score:
                  id: 4
                  userId: 1
                  username: player1
                  score: 150
                  mode: wall
                  timestamp: 1704067200000
        '401':
          description: Unauthorized - not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error: Not authenticated
        '400':
          description: Bad request - invalid score or mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    GameMode:
      type: string
      enum:
        - wall
        - pass
      description: |
        wall - Game ends when snake hits walls
        pass - Snake wraps around edges
      example: wall
    ScoreSubmission:
      type: object
      required:
        - score
        - mode
      properties:
        score:
          type: integer
          minimum: 0
          description: Game score
          example: 150
        mode:
          $ref: '#/components/schemas/GameMode'
    Score:
      type: object
      properties:
        id:
          type: integer
          description: Unique score identifier
          example: 1
        userId:
          type: integer
          description: ID of the user who achieved this score
          example: 1
        username:
          type: string
          description: Username of the player
          example: player1
        score:
          type: integer
          description: Game score
          example: 150
        mode:
          $ref: '#/components/schemas/GameMode'
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds when score was achieved
          example: 1704067200000
        rank:
          type: integer
          description: Rank position in leaderboard (only in leaderboard responses)
          example: 1
      required:
        - id
        - userId
        - username
        - score
        - mode
        - timestamp
    LeaderboardResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        leaderboard:
          type: array
          items:
            $ref: '#/components/schemas/Score'
      required:
        - success
        - leaderboard
    ScoreResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        score:
          $ref: '#/components/schemas/Score'
      required:
        - success
        - score

    # Players Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: Invalid email or password
      required:
        - success
        - error